LibELF+Utilities: Avoid truncating 64-bit values
This fixes displaying 64-bit addresses in readelf and also fixes
showing backtraces from core dumps on x86_64.

+++ Userland/Libraries/LibELF/CoreDump.h
@@ -50,8 +50,8 @@ struct [[gnu::packed]] ThreadInfo {
 
 struct [[gnu::packed]] MemoryRegionInfo {
     NotesEntryHeader header;
-    uint32_t region_start;
-    uint32_t region_end;
+    uint64_t region_start;
+    uint64_t region_end;
     uint16_t program_header_index;
     char region_name[]; // Null terminated
